Filter Node Overhaul: Add new filter component
