+++ packages/nodes-base/nodes/Filter/V2/FilterV2.node.ts
@@ -1,0 +1,111 @@
+import set from 'lodash/set';
+import type {
+	IExecuteFunctions,
+	INodeExecutionData,
+	INodeType,
+	INodeTypeBaseDescription,
+	INodeTypeDescription,
+} from 'n8n-workflow';
+
+export class FilterV2 implements INodeType {
+	description: INodeTypeDescription;
+
+	constructor(baseDescription: INodeTypeBaseDescription) {
+		this.description = {
+			...baseDescription,
+			version: 2,
+			defaults: {
+				name: 'Filter',
+				color: '#229eff',
+			},
+			inputs: ['main'],
+			outputs: ['main'],
+			outputNames: ['Kept', 'Discarded'],
+			properties: [
+				{
+					displayName: 'Conditions',
+					name: 'conditions',
+					placeholder: 'Add Condition',
+					type: 'filter',
+					default: {},
+					typeOptions: {
+						filter: {
+							caseSensitive: '={{!$parameter.options.ignoreCase}}',
+							typeValidation: '={{$parameter.options.looseTypeValidation ? "loose" : "strict"}}',
+						},
+					},
+				},
+				{
+					displayName: 'Options',
+					name: 'options',
+					type: 'collection',
+					placeholder: 'Add option',
+					default: {},
+					options: [
+						{
+							displayName: 'Ignore Case',
+							description: 'Whether to ignore letter case when evaluating conditions',
+							name: 'ignoreCase',
+							type: 'boolean',
+							default: true,
+						},
+						{
+							displayName: 'Less Strict Type Validation',
+							description: 'Whether to try casting value types based on the selected operator',
+							name: 'looseTypeValidation',
+							type: 'boolean',
+							default: true,
+						},
+					],
+				},
+			],
+		};
+	}
+
+	async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
+		const keptItems: INodeExecutionData[] = [];
+		const discardedItems: INodeExecutionData[] = [];
+
+		this.getInputData().forEach((item, itemIndex) => {
+			try {
+				const options = this.getNodeParameter('options', itemIndex) as {
+					ignoreCase?: boolean;
+					looseTypeValidation?: boolean;
+				};
+				let pass = false;
+				try {
+					pass = this.getNodeParameter('conditions', itemIndex, false, {
+						extractValue: true,
+					}) as boolean;
+				} catch (error) {
+					if (!options.looseTypeValidation) {
+						set(
+							error,
+							'description',
+							"Try to change the operator, switch ON the option 'Less Strict Type Validation', or change the type with an expression",
+						);
+					}
+					throw error;
+				}
+
+				if (item.pairedItem === undefined) {
+					item.pairedItem = { item: itemIndex };
+				}
+
+				if (pass) {
+					keptItems.push(item);
+				} else {
+					discardedItems.push(item);
+				}
+			} catch (error) {
+				if (this.continueOnFail()) {
+					discardedItems.push(item);
+				} else {
+					throw error;
+				}
+			}
+		});
+
+		return [keptItems, discardedItems];
+	}
+}
